Enable RSA verification
This should been as a development commit, since the created keys in this
commit *SHOULD NEVER EVER* be used to anything else than development and
demos. If someone intends to create a product, then of course, one needs
to update and replace the keys and remove the private key
(image_sign.pem) from the git itself.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>